+++ scripts/npmPublish.sh
@@ -82,6 +82,9 @@ function publishPackage {
     if [[ $VERSION != *"rc"* ]] && [[ $VERSION != *"next."* ]]; then
       echo "Tag $NAME@$VERSION with latest and next on npm"
       npm dist-tag add $NAME@$VERSION latest
+      M_TAG=`echo $VERSION | cut -d "." -f 1,2 | tr . -`
+      npm dist-tag add $NAME@$VERSION v$M_TAG
+      echo "Tag $NAME@$VERSION with latest, next and v${M_TAG} on npm"
     else
       echo "Release candidate $NAME@$VERSION NOT tagged as latest"
     fi